Move Loading Screen into its own component
To be able to use it on the Program page before the program info loads
